requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas==1.2.0
pandas-profiling==2.10.0
scikit-image==0.18.1
scikit-learn==0.22

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs==0.7.2
google-cloud-storage==1.36.0
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor
